backend/requirements.txt
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
torch
torchvision
opencv-python
Pillow
numpy

